old-tests/submission/W3C/domhtml/HTMLHRElement02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - HR</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-hr-noshade" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<hr align="center" noshade="noShade" size="5" width="400">



<script>
var domtest = async_test("The noShade IDL attribute of the hr element must reflect the element's noshade content attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vnoshade;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("hr");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vnoshade = testNode.noShade;

      assert_true(vnoshade);

    domtest.done();
}));

</script>
<p>The noShade IDL attribute of the hr element must reflect the element's noshade content attribute.</p>
<div id="log"></div>
</body></html>
